+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Added
 + Added support for curves
++ Added minimal export option, which skips textures for faster export
 
 ### Fixed
 * Improved materials rgb_curves node support